Consolidate StickyHeaderBehavior and QuickReturnHeaderBehavior with new abstract HeaderBehaviorBase helper class
Is responsible for finding needed controls and properties for setting up composition animation. 90% of these two classes were doing the same thing, so now only the logic related to setting up/manipulating the composition animation is in each subclass.
Going to investigate if FadeHeaderBehavior is similar enough to reuse as well.
